XUI: - Reflection type safe of casting and fix issue when reflection parse not find reflection object at finalize stage. - Reflection runtime: Variant GetValue method type safety (partial) - Added missing virtual destructors - Fix MenuItem default control template - Fix Popup pointer placement when use GameUI mode

+++ src/IrrLicht/CIrrDeviceStub.h
@@ -58,7 +58,7 @@ namespace irr
 		//! destructor
 		virtual ~CIrrDeviceStub();
 
-        virtual int getHandle() { return 0; }
+        virtual intptr_t getHandle() { return 0; }
 
 		//! returns the video driver
 		virtual video::IVideoDriver* getVideoDriver();
